rename: token, tokenProvider 패키지 이동
