+++ src/main/java/back/ecommerce/auth/interceptor/JwtAuthenticationInterceptor.java
@@ -1,0 +1,2 @@
+package back.ecommerce.auth.interceptor;public class JwtAuthenticationInterceptor {
+}
