+++ src/main/java/back/ecommerce/auth/token/Token.java
@@ -1,4 +1,4 @@
-package back.ecommerce.auth;
+package back.ecommerce.auth.token;
 
 import lombok.AllArgsConstructor;
 import lombok.Getter;
