+++ src/main/java/back/ecommerce/exception/TokenIsEmptyException.java
@@ -1,0 +1,2 @@
+package back.ecommerce.exception;public class TokenIsEmptyException {
+}
