+++ src/main/java/back/ecommerce/service/AuthService.java
@@ -2,8 +2,8 @@ package back.ecommerce.service;
 
 import org.springframework.security.crypto.password.PasswordEncoder;
 import org.springframework.stereotype.Service;
-import back.ecommerce.auth.Token;
-import back.ecommerce.auth.TokenProvider;
+import back.ecommerce.auth.token.Token;
+import back.ecommerce.auth.token.TokenProvider;
 import back.ecommerce.domain.User;
 import back.ecommerce.dto.response.token.TokenResponseDto;
 import back.ecommerce.exception.PasswordNotMatchedException;
