+++ src/test/java/back/ecommerce/auth/TokenProviderTest.java
@@ -7,6 +7,8 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.beans.factory.annotation.Value;
 import org.springframework.boot.test.context.SpringBootTest;
 
+import back.ecommerce.auth.token.Token;
+import back.ecommerce.auth.token.TokenProvider;
 import io.jsonwebtoken.Jwts;
 
 @SpringBootTest(classes = TokenProvider.class)
